+++ src/MicroElements.Metadata.NewtonsoftJson/NewtonsoftJson/ConfigurationExtensions.cs
@@ -1,0 +1,24 @@
+﻿// Copyright (c) MicroElements. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Newtonsoft.Json;
+
+namespace MicroElements.Metadata.NewtonsoftJson
+{
+    /// <summary>
+    /// Configuration extensions.
+    /// </summary>
+    public static class ConfigurationExtensions
+    {
+        /// <summary>
+        /// Configures <see cref="JsonSerializerSettings"/> to serialize <see cref="IPropertyContainer"/>.
+        /// </summary>
+        /// <param name="options"><see cref="JsonSerializerSettings"/> instance.</param>
+        /// <returns>The same options.</returns>
+        public static JsonSerializerSettings ConfigureForMetadata(this JsonSerializerSettings options)
+        {
+            options.Converters.Add(new PropertyContainerConverter());
+            return options;
+        }
+    }
+}