+++ src/MicroElements.Metadata.SystemTextJson/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace MicroElements.Metadata.SystemTextJson
+{
+    public class Class1
+    {
+    }
+}